% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fitTMB}
\alias{fitTMB}
\title{Fit TMB model}
\usage{
fitTMB(tmbArg, verbose = FALSE, control = list())
}
\arguments{
\item{tmbArg}{An object from \code{getArg()} containing arguments for \code{TMB::MakeADFun()}}

\item{verbose}{Logical. If \code{TRUE}, print maximum gradient \code{mgc} components while fitting.}

\item{control}{Optional control arguments for \code{nlminb()}. Supply as \code{list()}.}
}
\description{
Fit TMB model
}
