% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{fixef.snaTMB}
\alias{fixef.snaTMB}
\alias{fixef}
\title{Get fixed effects}
\usage{
\method{fixef}{snaTMB}(object, ...)
}
\arguments{
\item{object}{Any fitted model object from which fixed effects estimates can be extracted.}

\item{\dots}{Additional arguments. Currently none.}
}
\description{
Get fixed effects
}
