% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{fixef.snglmm}
\alias{fixef.snglmm}
\alias{fixef}
\title{Extract Fixed Effects}
\usage{
\method{fixef}{snglmm}(object, ...)
}
\arguments{
\item{object}{any fitted model object from which fixed effects
   estimates can be extracted.}

\item{...}{some methods for this generic function require additional
   arguments.}
}
\description{
This function is generic; method functions can be written to handle specific classes of objects.
}
