% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{format_fixef}
\alias{format_fixef}
\title{Format fixed effects}
\usage{
format_fixef(sdr, tmbArg, DF = NULL)
}
\arguments{
\item{sdr}{Output from \code{TMB::sdreport()}}

\item{tmbArg}{Output from \code{snaTMB::getArg()}}

\item{DF}{Degree of freedom for Student-t distribution}
}
\description{
Format fixed effects
}
