% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{format_ranef}
\alias{format_ranef}
\title{Format random effects}
\usage{
format_ranef(sdr, tmbArg)
}
\arguments{
\item{sdr}{Output from \code{TMB::sdreport()}}

\item{tmbArg}{Output from \code{snaTMB::getArg()}}
}
\description{
Format random effects
}
