% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{format_spatial}
\alias{format_spatial}
\title{Format spatial effects}
\usage{
format_spatial(sdr, tmbArg, full = FALSE)
}
\arguments{
\item{sdr}{Output from \code{TMB::sdreport()}}

\item{tmbArg}{Output from \code{snaTMB::getArg()}}

\item{full}{Logical. Full report with SEs if \code{TRUE}.}
}
\description{
Format spatial effects
}
