% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_arg}
\alias{get_arg}
\title{Get arguments for MakeADfun}
\usage{
get_arg(
  formula,
  data,
  family,
  sigma_in = TRUE,
  spatial = NULL,
  D = NULL,
  W = NULL
)
}
\arguments{
\item{formula}{An object of class "formula" (or one that can be coerced to that class): a symbolic description of the model to be fitted.}

\item{data}{An optional data frame, list or environment (or object coercible by as.data.frame to a data frame) containing the variables in the model. If not found in data, the variables are taken from environment(formula).}

\item{family}{Family assumed for an error distribution. Default \code{gaussian()}.}

\item{sigma_in}{An optional argument for on/off non-spatial normal error variance. If \code{TRUE}, a non-spatial normal error variance will be included in a spatial model. Default \code{TRUE}.}

\item{spatial}{Model type used for a spatial model (either \code{"exp"} or \code{"gaussian"}). Disabled by default.}

\item{D}{An optional distance matrix for a spatial model.}

\item{W}{An optional spatial weight matrix for a spatial model.}
}
\description{
Get arguments for MakeADfun
}
