% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getArg}
\alias{getArg}
\title{Get arguments for MakeADfun}
\usage{
getArg(formula, data, family, spatial = NULL, D = NULL, W = NULL)
}
\arguments{
\item{formula}{an object of class "formula" (or one that can be coerced to that class): a symbolic description of the model to be fitted.}

\item{data}{a data frame containing the variables in the model.}

\item{family}{a probability distribution used as an error distribution}

\item{D}{distance matrix}

\item{W}{spatial weight matrix}
}
\description{
Get arguments for MakeADfun
}
\author{
Akira Terui, \email{hanabi0111@gmail.com}
}
