% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snglmm.R
\name{kriging}
\alias{kriging}
\title{Perform Regression/Universal Kriging}
\usage{
kriging(object, newdata, cD, cW)
}
\arguments{
\item{object}{Object class \code{snglmm}}

\item{newdata}{A required data frame containing predictor values at new sites.}

\item{cD}{A required cross-distance matrix for a spatial model.}

\item{cW}{An optional cross-weight matrix for a spatial model.}
}
\description{
Perform Regression/Universal Kriging
}
