% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.snglmm}
\alias{print.snglmm}
\alias{print}
\title{Print Values}
\usage{
\method{print}{snglmm}(x, digits = max(3L, getOption("digits") - 3L), ...)
}
\arguments{
\item{x}{an object used to select a method.}

\item{digits}{minimal number of \emph{significant} digits, see
    \code{\link[base]{print.default}}.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
\code{print} prints its argument and returns it invisibly (via \code{\link{invisible}}(x)). It is a generic function which means that new printing methods can be easily added for new \code{\link{class}}es.
}
