% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_semtmb.R
\name{semtmb}
\alias{semtmb}
\title{Fit a spatial error model (SEM) with TMB}
\usage{
semtmb(formula, data)
}
\arguments{
\item{formula}{an object of class "formula" (or one that can be coerced to that class): a symbolic description of the model to be fitted.}

\item{data}{an optional data frame, list or environment (or object coercible by as.data.frame to a data frame) containing the variables in the model. If not found in data, the variables are taken from environment(formula).}
}
\description{
Fit a spatial error model (SEM) with TMB
}
\author{
Akira Terui, \email{hanabi0111@gmail.com}
}
