% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snaTMB-package.R
\docType{package}
\name{snaTMB-package}
\alias{snaTMB}
\alias{snaTMB-package}
\title{snaTMB: Stream Network Analysis with Template Model Builder}
\description{
snaTMB provides a set of function to perform spatial analysis in stream networks.
}
\keyword{internal}
