% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snaTMB.R
\name{snaTMB}
\alias{snaTMB}
\title{Fit a spatial model with TMB}
\usage{
snaTMB(
  formula,
  data,
  family = gaussian(),
  spatial,
  D,
  W,
  verbose = FALSE,
  inits,
  control = list()
)
}
\arguments{
\item{formula}{An object of class "formula" (or one that can be coerced to that class): a symbolic description of the model to be fitted.}

\item{data}{An optional data frame, list or environment (or object coercible by as.data.frame to a data frame) containing the variables in the model. If not found in data, the variables are taken from environment(formula).}

\item{family}{Family assumed for an error distribution. Default \code{gaussian()}.}

\item{spatial}{Model type used for a spatial model (either \code{"exp"} or \code{"gaussian"}). Disabled by default.}

\item{D}{An optional distance matrix for a spatial model.}

\item{W}{An optional spatial weight matrix for a spatial model.}

\item{verbose}{Logical. If \code{TRUE}, print maximum gradient \code{mgc} components while fitting.}

\item{inits}{A list of initial parameter values.}

\item{control}{Optional arguments passed to \code{\link{nlminb}}.}
}
\description{
Fit a spatial model with TMB
}
\author{
Akira Terui, \email{hanabi0111@gmail.com}
}
