% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.snaTMB}
\alias{summary.snaTMB}
\alias{summary}
\title{Summarize snaTMB output}
\usage{
\method{summary}{snaTMB}(object, digits = max(3L, getOption("digits") - 3L), ...)
}
\arguments{
\item{object}{An object for which a summary is desired.}

\item{\dots}{Additional arguments affecting the summary produced.}
}
\description{
Summarize snaTMB output
}
