% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.snglmm}
\alias{summary.snglmm}
\alias{summary}
\title{Object Summaries}
\usage{
\method{summary}{snglmm}(object, digits = max(3L, getOption("digits") - 3L), ...)
}
\arguments{
\item{object}{an object for which a summary is desired.}

\item{digits}{integer, used for number formatting with
    \code{\link[base]{signif}()} (for \code{summary.default}) or
    \code{\link[base]{format}()} (for \code{summary.data.frame}).  In
    \code{summary.default}, if not specified (i.e.,
    \code{\link[base]{missing}(.)}), \code{signif()} will \emph{not} be called
    anymore (since \R >= 3.4.0, where the default has been changed to
    only round in the \code{print} and \code{format} methods).}

\item{...}{additional arguments affecting the summary produced.}
}
\description{
\code{summary} is a generic function used to produce result summaries of the results of various model fitting functions. The function invokes particular \code{\link{methods}} which depend on the \code{\link{class}} of the first argument.
}
