% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{vcov.snaTMB}
\alias{vcov.snaTMB}
\alias{vcov}
\title{Get variance-covariance for fixed effects}
\usage{
\method{vcov}{snaTMB}(object, ...)
}
\arguments{
\item{object}{A fitted model object, typically.}

\item{\dots}{Additional arguments. Currently none.}
}
\description{
Get variance-covariance for fixed effects
}
