% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{vcov.snglmm}
\alias{vcov.snglmm}
\alias{vcov}
\title{Calculate Variance-Covariance Matrix for a Fitted Model Object}
\usage{
\method{vcov}{snglmm}(object, ...)
}
\arguments{
\item{object}{
    a fitted model object, typically.  Sometimes also a
    \code{\link{summary}()} object of such a fitted model.
  }

\item{...}{
    additional arguments for method functions.  For the
    \code{\link[stats]{glm}} method this can be used to pass a
    \code{dispersion} parameter.}
}
\description{
Returns the variance-covariance matrix of the main parameters of a fitted model object. The “main” parameters of model correspond to those returned by \code{\link{coef}}, and typically do not contain a nuisance scale parameter (\code{\link{sigma}}).
}
