% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{format_fixef}
\alias{format_fixef}
\title{Format fixed effects}
\usage{
format_fixef(sdr, tmb_arg)
}
\arguments{
\item{sdr}{Output from \code{TMB::sdreport}}

\item{tmb_arg}{Output from \code{snaTMB::get_arg}}
}
\description{
Format fixed effects
}
