% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{format_ranef}
\alias{format_ranef}
\title{Format random effects}
\usage{
format_ranef(fit, sdr, tmb_arg, full = FALSE)
}
\arguments{
\item{fit}{Output from \code{snaTMB::snaTMB}}

\item{sdr}{Output from \code{TMB::sdreport}}

\item{tmb_arg}{Output from \code{snaTMB::get_arg}}

\item{full}{Logical. If TRUE, standard errors are printed.}
}
\description{
Format random effects
}
