% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{format_spatial}
\alias{format_spatial}
\title{Format spatial effects}
\usage{
format_spatial(sdr, tmb_arg, full = FALSE)
}
\arguments{
\item{sdr}{Output from \code{TMB::sdreport}}

\item{tmb_arg}{Output from \code{snglmm::get_arg}}

\item{full}{Logical. If TRUE, standard errors are printed.}
}
\description{
Format spatial effects
}
