% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.snaTMB}
\alias{print.snaTMB}
\alias{print}
\title{Print snaTMB output}
\usage{
\method{print}{snaTMB}(x, digits = max(3L, getOption("digits") - 3L), ...)
}
\arguments{
\item{x}{Object class snaTMB}

\item{digits}{Minimal number of significant digits, see \code{\link{print.default}}.}

\item{\dots}{Additional arguments for \code{\link{print.default}}.}
}
\description{
Print snaTMB output
}
